Allow initialization of WCSAxes from transform/coord_meta
